refactor(structured-argumentation): migrate to code mode architecture
Separate concerns into core, codemode, and mcp layers for better maintainability and programmatic usage. Core layer contains pure business logic, codemode provides a typed TypeScript API, and mcp handles MCP protocol adaptation. This enables both MCP tool usage and direct library integration.

- Extract types and business logic to src/core/
- Create StructuredArgumentation class in src/codemode/ for programmatic API
- Implement MCP adapter in src/mcp/ that uses codemode API
- Update build configuration to exclude tests from production builds
- Refactor tests to directly test codemode API
- Update documentation with architecture overview and usage examples
