+++ .trae/documents/plan_20260213_002440.md
@@ -1,0 +1,45 @@
+# Migration Plan: Structured Argumentation to Code Mode
+
+I will migrate the `structured-argumentation` MCP server to the Code Mode architecture, separating concerns into Core (logic), Code Mode (API), and MCP (transport) layers.
+
+## 1. Directory Structure Setup
+- Create `src/core`, `src/codemode`, and `src/mcp` directories.
+
+## 2. Core Layer (`src/core/`)
+- **`src/core/types.ts`**: Extract `ArgumentData`, `ArgumentType`, and internal interfaces.
+- **`src/core/manager.ts`**:
+  - Implement `ArgumentationManager` class.
+  - Move state (`argumentHistory`, `relationshipGraph`) and logic (`validateArgumentData`, `updateRelationshipGraph`).
+  - Move `formatArgument` and `getArgumentGraphSummary` logic here (or to a separate formatter).
+  - Ensure pure logic without MCP SDK dependencies.
+
+## 3. Code Mode Layer (`src/codemode/`)
+- **`src/codemode/index.ts`**:
+  - Create `StructuredArgumentation` class.
+  - Expose typed methods: `processArgument(input: ArgumentData): Promise<ArgumentResult>`.
+  - Handle input validation and delegate to `ArgumentationManager`.
+
+## 4. MCP Adapter Layer (`src/mcp/`)
+- **`src/mcp/tools.ts`**:
+  - Define `STRUCTURED_ARGUMENTATION_TOOL` schema (moved from `index.ts`).
+- **`src/mcp/index.ts`**:
+  - Create the MCP server instance.
+  - Instantiate `StructuredArgumentation`.
+  - Map MCP tool calls to Code Mode API methods.
+
+## 5. Entry Point & Config
+- **`src/index.ts`**:
+  - Export `StructuredArgumentation` from `codemode`.
+  - Conditional execution for running as an MCP server (maintain `import.meta.main` logic).
+- **`package.json`**:
+  - Update `exports` to point to the new structure.
+  - Ensure `types` point to the correct declaration files.
+
+## 6. Testing & Documentation
+- **Refactor Tests**: Update `src/index.test.ts` to test `StructuredArgumentation` class directly (Code Mode testing).
+- **Documentation**: Update `README.md` to document the new Programmatic API (Code Mode) and the MCP Tool usage.
+- **Migration Log**: Create `MIGRATION_LOG.md` to document the changes.
+
+## 7. Verification
+- Run `bun test` to ensure no regression.
+- Verify the server builds and runs.
