+++ src/structured-argumentation/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@wemake.cx/structured-argumentation",
-  "version": "0.3.0",
+  "version": "0.4.0",
   "description": "MCP server for diagrammatic thinking and spatial representation",
   "license": "MIT",
   "main": "./dist/index.js",
@@ -34,7 +34,7 @@
     "dist"
   ],
   "scripts": {
-    "build": "bun build ./src/index.ts --outdir='./dist' --target='bun' && tsc",
+    "build": "bun build ./src/index.ts --outdir='./dist' --target='bun' && tsc -p tsconfig.build.json",
     "start": "bun run ./dist/index.js",
     "test": "bun test",
     "publish": "bun publish"
