+++ src/structured-argumentation/src/codemode/index.ts
@@ -1,0 +1,22 @@
+import { ArgumentationManager } from "../core/manager.js";
+import { ArgumentAnalysis, ArgumentData, ArgumentType } from "../core/types.js";
+
+export class StructuredArgumentation {
+  private manager: ArgumentationManager;
+
+  constructor() {
+    this.manager = new ArgumentationManager();
+  }
+
+  /**
+   * Processes an argument input, updates the internal state, and returns analysis.
+   *
+   * @param input - The argument data to process.
+   * @returns The analysis of the argument.
+   */
+  public async processArgument(input: unknown): Promise<ArgumentAnalysis> {
+    return this.manager.processArgument(input);
+  }
+}
+
+export type { ArgumentAnalysis, ArgumentData, ArgumentType };
