+++ src/structured-argumentation/src/mcp/tools.ts
@@ -1,0 +1,114 @@
+import { Tool } from "@modelcontextprotocol/sdk/types.js";
+
+export const STRUCTURED_ARGUMENTATION_TOOL: Tool = {
+  name: "structuredArgumentation",
+  description: `A detailed tool for systematic dialectical reasoning and argument analysis.
+This tool helps analyze complex questions through formal argumentation structures.
+It facilitates the creation, critique, and synthesis of competing arguments.
+
+When to use this tool:
+- Evaluating competing perspectives and claims
+- Analyzing complex ethical dilemmas
+- Assessing policy proposals with multiple stakeholders
+- Exploring scientific hypotheses and counter-arguments
+
+Key features:
+- Break down arguments into claims, premises, and conclusions
+- Track relationships between arguments
+- Represent objections and rebuttals
+- Facilitate dialectical progression through thesis-antithesis-synthesis
+- Evaluate argument strengths and weaknesses
+- Visualize argument structures
+
+Parameters explained:
+- claim: The central proposition being argued
+- premises: Supporting evidence or assumptions
+- conclusion: The logical consequence of accepting the claim
+- argumentType: Whether this is a thesis, antithesis, synthesis, objection, or rebuttal
+- confidence: Your confidence level in this argument (0.0-1.0)
+- respondsTo: ID of an argument this directly responds to
+- supports/contradicts: IDs of arguments this supports or contradicts
+- strengths/weaknesses: Notable strong or weak points of the argument
+- nextArgumentNeeded: Whether another argument is needed in the dialectic`,
+
+  inputSchema: {
+    type: "object",
+    properties: {
+      claim: {
+        type: "string",
+        description: "The central proposition being argued"
+      },
+      premises: {
+        type: "array",
+        description: "Supporting evidence or assumptions",
+        items: {
+          type: "string"
+        }
+      },
+      conclusion: {
+        type: "string",
+        description: "The logical consequence of accepting the claim"
+      },
+      argumentId: {
+        type: "string",
+        description: "Optional unique identifier for this argument"
+      },
+      argumentType: {
+        type: "string",
+        enum: ["thesis", "antithesis", "synthesis", "objection", "rebuttal"],
+        description: "The type of argument being presented"
+      },
+      confidence: {
+        type: "number",
+        description: "Confidence level in this argument (0.0-1.0)",
+        minimum: 0,
+        maximum: 1
+      },
+      respondsTo: {
+        type: "string",
+        description: "ID of the argument this directly responds to"
+      },
+      supports: {
+        type: "array",
+        description: "IDs of arguments this supports",
+        items: {
+          type: "string"
+        }
+      },
+      contradicts: {
+        type: "array",
+        description: "IDs of arguments this contradicts",
+        items: {
+          type: "string"
+        }
+      },
+      strengths: {
+        type: "array",
+        description: "Notable strong points of the argument",
+        items: {
+          type: "string"
+        }
+      },
+      weaknesses: {
+        type: "array",
+        description: "Notable weak points of the argument",
+        items: {
+          type: "string"
+        }
+      },
+      nextArgumentNeeded: {
+        type: "boolean",
+        description: "Whether another argument is needed in the dialectic"
+      },
+      suggestedNextTypes: {
+        type: "array",
+        description: "Suggested types for the next argument",
+        items: {
+          type: "string",
+          enum: ["thesis", "antithesis", "synthesis", "objection", "rebuttal"]
+        }
+      }
+    },
+    required: ["claim", "premises", "conclusion", "argumentType", "confidence", "nextArgumentNeeded"]
+  }
+};
